execution/_template.py
#!/usr/bin/env python3
"""
Script: _template.py
Version: 1.0.0
Purpose: [Brief description of what this script does]

Usage:
    python execution/_template.py --input <value>

Notes:
    - This is a template for execution scripts
    - All execution scripts should be deterministic and testable
    - Load secrets from .env, not hardcoded
"""

import argparse
import json
import os
import sys
from pathlib import Path

from dotenv import load_dotenv

# =============================================================================
# CONFIGURATION
# =============================================================================
load_dotenv()

# Example: API_KEY = os.getenv("API_KEY")


# =============================================================================
# CORE FUNCTIONS
# =============================================================================
def main(input_value: str) -> dict:
    """
    Main execution logic.
    
    Args:
        input_value: The primary input for this script.
        
    Returns:
        A dictionary containing the results.
    """
    result = {
        "status": "success",
        "input": input_value,
        "output": None,
    }
    
    # TODO: Implement core logic here
    
    return result


# =============================================================================
# CLI ENTRYPOINT
# =============================================================================
if __name__ == "__main__":
    parser = argparse.ArgumentParser(description="Template execution script")
    parser.add_argument("--input", required=True, help="Primary input value")
    parser.add_argument("--output", default=".tmp/output.json", help="Output file path")
    
    args = parser.parse_args()
    
    # Ensure .tmp directory exists
    Path(".tmp").mkdir(exist_ok=True)
    
    # Execute
    result = main(args.input)
    
    # Write output
    with open(args.output, "w") as f:
        json.dump(result, f, indent=2)
    
    print(f"✓ Output written to {args.output}")
    sys.exit(0 if result["status"] == "success" else 1)
