requirements.txt
# =============================================================================
# Python Dependencies — 3-Layer Agent Architecture
# =============================================================================
# Install: pip install -r requirements.txt
# =============================================================================

# --- Core ---
python-dotenv>=1.0.0

# --- Google APIs (uncomment as needed) ---
# google-auth>=2.0.0
# google-auth-oauthlib>=1.0.0
# google-api-python-client>=2.0.0
# gspread>=5.0.0

# --- HTTP & Scraping (uncomment as needed) ---
requests>=2.28.0
# httpx>=0.24.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
webdriver-manager>=4.0.0

# --- Data Processing (uncomment as needed) ---
# pandas>=2.0.0
# pydantic>=2.0.0

# --- Statistics & Visualization ---
numpy>=1.26.0
scipy>=1.11.0
